+++ packages/ui-services/src/Route/Params/PurchaseParams.ts
@@ -1,0 +1,4 @@
+export type PurchaseParams = {
+  plan: string
+  period: string
+}