+++ packages/ui-services/src/Route/Params/SettingsParams.ts
@@ -1,0 +1,5 @@
+import { PreferenceId } from '../../Preferences/PreferenceId'
+
+export type SettingsParams = {
+  panel: PreferenceId
+}