+++ packages/ui-services/src/Route/Params/SubscriptionInviteParams.ts
@@ -1,0 +1,5 @@
+import { Uuid } from '@standardnotes/common'
+
+export type SubscriptionInviteParams = {
+  inviteUuid: Uuid
+}