+++ packages/ui-services/src/Route/RootQueryParam.ts
@@ -1,0 +1,6 @@
+export enum RootQueryParam {
+  Purchase = 'purchase',
+  Settings = 'settings',
+  DemoToken = 'demo-token',
+  AcceptSubscriptionInvite = 'accept-subscription-invite',
+}